Moved database config file up two levels.
